+++ packages/blocks-core/js/libs/wordpress/audio/test/functionality.blocks-1.e2e.cy.js
@@ -28,7 +28,7 @@ describe('Audio Block', () => {
 		// Block supported is active
 		cy.get('.blockera-extension-block-card').should('be.visible');
 
-		cy.checkBlockCardItems(['normal', 'hover']);
+		cy.checkBlockCardItems(['normal', 'hover', 'elements/caption']);
 
 		//
 		// 1.0. Block Styles